어드민 총 업체 통계 완성

+++ tracky-web/src/main/java/kernel360/trackyweb/admin/statistic/application/dto/response/HourlyGraphResponse.java
@@ -1,25 +1,27 @@
 package kernel360.trackyweb.admin.statistic.application.dto.response;
 
 
+import lombok.extern.slf4j.Slf4j;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
+
 public record HourlyGraphResponse(
         int hour,
-        int driveCount
+        long driveCount
 ) {
     public static List<HourlyGraphResponse> toResultList(List<HourlyGraphResponse> latestData) {
-        Map<Integer, Integer> hourToCount = latestData.stream()
+        Map<Integer, Long> hourToCount = latestData.stream()
                 .collect(Collectors.toMap(HourlyGraphResponse::hour, HourlyGraphResponse::driveCount));
 
         List<HourlyGraphResponse> fullList = new ArrayList<>();
         for (int i = 0; i < 24; i++) {
-            int count = hourToCount.getOrDefault(i, 0);
+            long count = hourToCount.getOrDefault(i, 0L);
             fullList.add(new HourlyGraphResponse(i, count));
         }
-
         return fullList;
     }
 }
